yolov3-torch/requirements.txt
numpy~=1.19.4
pillow~=8.0.1
torch>=1.0
torchvision~=0.8.2

tqdm~=4.55.1
pandas~=1.1.5
sacred~=0.8.2
matplotlib~=3.3.3
pykitti~=0.3.1
opencv-python~=4.5.1.48